% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{actual_vapor_pressure}
\alias{actual_vapor_pressure}
\title{Actual Vapor Pressure}
\usage{
actual_vapor_pressure(Tmin, Tmax, RH_morning, RH_evening)
}
\arguments{
\item{Tmin}{Minimum air temperature (degrees Celsius).}

\item{Tmax}{Maximum air temperature (degrees Celsius).}

\item{RH_morning}{Morning relative humidity (percent).}

\item{RH_evening}{Evening relative humidity (percent).}
}
\value{
Actual vapor pressure (kPa).
}
\description{
Computes actual vapor pressure from relative humidity and temperature.
}
\examples{
actual_vapor_pressure(15, 30, 90, 40)
}
\references{
Allen et al. (1998), FAO-56, Eq. 17.
}
