% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_mae}
\alias{calc_mae}
\title{Mean Absolute Error (MAE)}
\usage{
calc_mae(observed, predicted)
}
\arguments{
\item{observed}{Numeric vector of observed (reference) values.}

\item{predicted}{Numeric vector of predicted (model) values.}
}
\value{
MAE value (numeric scalar).
}
\description{
Mean Absolute Error (MAE)
}
\examples{
calc_mae(c(3.5, 4.2, 5.1), c(3.3, 4.5, 4.9))
}
