% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_mse}
\alias{calc_mse}
\title{Mean Squared Error (MSE)}
\usage{
calc_mse(observed, predicted)
}
\arguments{
\item{observed}{Numeric vector of observed (reference) values.}

\item{predicted}{Numeric vector of predicted (model) values.}
}
\value{
MSE value (numeric scalar).
}
\description{
Mean Squared Error (MSE)
}
\examples{
calc_mse(c(3.5, 4.2, 5.1), c(3.3, 4.5, 4.9))
}
