% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_r}
\alias{calc_r}
\title{Pearson Correlation Coefficient (r)}
\usage{
calc_r(observed, predicted)
}
\arguments{
\item{observed}{Numeric vector of observed (reference) values.}

\item{predicted}{Numeric vector of predicted (model) values.}
}
\value{
Pearson correlation coefficient (numeric scalar).
}
\description{
Pearson Correlation Coefficient (r)
}
\examples{
calc_r(c(3.5, 4.2, 5.1), c(3.3, 4.5, 4.9))
}
