% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_rmse}
\alias{calc_rmse}
\title{Root Mean Squared Error (RMSE)}
\usage{
calc_rmse(observed, predicted)
}
\arguments{
\item{observed}{Numeric vector of observed (reference) values.}

\item{predicted}{Numeric vector of predicted (model) values.}
}
\value{
RMSE value (numeric scalar).
}
\description{
Root Mean Squared Error (RMSE)
}
\examples{
calc_rmse(c(3.5, 4.2, 5.1), c(3.3, 4.5, 4.9))
}
