% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{daylight_hours}
\alias{daylight_hours}
\title{Daylight Hours}
\usage{
daylight_hours(J, lat)
}
\arguments{
\item{J}{Day of the year (1-366).}

\item{lat}{Latitude (degrees, negative for Southern Hemisphere).}
}
\value{
Maximum daylight hours.
}
\description{
Computes the maximum possible daylight hours for a given day and latitude.
}
\examples{
daylight_hours(172, 25.43)
}
\references{
Allen et al. (1998), FAO-56, Eq. 34.
}
