% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dew_point_temperature}
\alias{dew_point_temperature}
\title{Dew Point Temperature}
\usage{
dew_point_temperature(ea)
}
\arguments{
\item{ea}{Actual vapor pressure (kPa).}
}
\value{
Dew point temperature (degrees Celsius).
}
\description{
Estimates dew point temperature from actual vapor pressure.
}
\examples{
ea <- actual_vapor_pressure(15, 30, 90, 40)
dew_point_temperature(ea)
}
