% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0_temperature.R
\name{et0_baier_robertson}
\alias{et0_baier_robertson}
\title{Baier-Robertson ET0 Model}
\usage{
et0_baier_robertson(Tmin, Tmax, J, lat)
}
\arguments{
\item{Tmin}{Minimum air temperature (degrees Celsius).}

\item{Tmax}{Maximum air temperature (degrees Celsius).}

\item{J}{Day of the year (1-366).}

\item{lat}{Latitude (degrees).}
}
\value{
Daily ET0 (mm/day).
}
\description{
Estimates daily ET0 using the Baier and Robertson (1965) method, a
regression-based equation using extraterrestrial radiation, maximum
temperature and temperature range.
}
\examples{
et0_baier_robertson(15, 30, 172, 25.43)
}
\references{
Baier, W. & Robertson, G.W. (1965). Estimation of latent
  evaporation from simple weather observations. Canadian Journal of Plant
  Science, 45(3), 276-284.
}
\seealso{
Other temperature models: 
\code{\link{et0_berti}()},
\code{\link{et0_blaney_criddle}()},
\code{\link{et0_dorji}()},
\code{\link{et0_droogers_allen}()},
\code{\link{et0_hargreaves_samani}()},
\code{\link{et0_linacre}()},
\code{\link{et0_schendel}()},
\code{\link{et0_tabari_talee1}()},
\code{\link{et0_tabari_talee2}()}
}
\concept{temperature models}
