% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0_temperature.R
\name{et0_berti}
\alias{et0_berti}
\title{Berti et al. ET0 Model}
\usage{
et0_berti(Tmin, Tmax, J, lat)
}
\arguments{
\item{Tmin}{Minimum air temperature (degrees Celsius).}

\item{Tmax}{Maximum air temperature (degrees Celsius).}

\item{J}{Day of the year (1-366).}

\item{lat}{Latitude (degrees).}
}
\value{
Daily ET0 (mm/day).
}
\description{
Estimates daily ET0 using the Berti et al. (2014) method, a modified
Hargreaves formula using the temperature range raised to the power 0.517.
}
\examples{
et0_berti(15, 30, 172, 25.43)
}
\references{
Berti, A., Tardivo, G., Chiaudani, A., Rech, F., & Borin, M.
  (2014). Assessing reference evapotranspiration by the Hargreaves
  method in north-eastern Italy. Agricultural Water Management, 140, 20-25.
}
\seealso{
Other temperature models: 
\code{\link{et0_baier_robertson}()},
\code{\link{et0_blaney_criddle}()},
\code{\link{et0_dorji}()},
\code{\link{et0_droogers_allen}()},
\code{\link{et0_hargreaves_samani}()},
\code{\link{et0_linacre}()},
\code{\link{et0_schendel}()},
\code{\link{et0_tabari_talee1}()},
\code{\link{et0_tabari_talee2}()}
}
\concept{temperature models}
