% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0_temperature.R
\name{et0_blaney_criddle}
\alias{et0_blaney_criddle}
\title{Blaney-Criddle ET0 Model}
\usage{
et0_blaney_criddle(Tmean)
}
\arguments{
\item{Tmean}{Mean air temperature (degrees Celsius).}
}
\value{
Daily ET0 (mm/day).
}
\description{
Estimates daily ET0 using the Blaney-Criddle (1950) method, which relies
only on mean air temperature.
}
\examples{
et0_blaney_criddle(25)
}
\references{
Blaney, H.F. & Criddle, W.D. (1950). Determining water
  requirements in irrigated areas from climatological and irrigation data.
  USDA Soil Conservation Service, SCS-TP 96.
}
\seealso{
Other temperature models: 
\code{\link{et0_baier_robertson}()},
\code{\link{et0_berti}()},
\code{\link{et0_dorji}()},
\code{\link{et0_droogers_allen}()},
\code{\link{et0_hargreaves_samani}()},
\code{\link{et0_linacre}()},
\code{\link{et0_schendel}()},
\code{\link{et0_tabari_talee1}()},
\code{\link{et0_tabari_talee2}()}
}
\concept{temperature models}
