% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0_temperature.R
\name{et0_dorji}
\alias{et0_dorji}
\title{Dorji et al. ET0 Model}
\usage{
et0_dorji(Tmin, Tmax, J, lat)
}
\arguments{
\item{Tmin}{Minimum air temperature (degrees Celsius).}

\item{Tmax}{Maximum air temperature (degrees Celsius).}

\item{J}{Day of the year (1-366).}

\item{lat}{Latitude (degrees).}
}
\value{
Daily ET0 (mm/day).
}
\description{
Estimates daily ET0 using the Dorji et al. (2016) method, a modified
Hargreaves formula using the temperature range raised to the power 0.296.
}
\examples{
et0_dorji(15, 30, 172, 25.43)
}
\references{
Dorji, U., Olesen, J.E., & Seidenkrantz, M.S. (2016). Water
  balance in the complex mountainous terrain of Bhutan and linkages to land
  use. Journal of Hydrology: Regional Studies, 7, 55-68.
}
\seealso{
Other temperature models: 
\code{\link{et0_baier_robertson}()},
\code{\link{et0_berti}()},
\code{\link{et0_blaney_criddle}()},
\code{\link{et0_droogers_allen}()},
\code{\link{et0_hargreaves_samani}()},
\code{\link{et0_linacre}()},
\code{\link{et0_schendel}()},
\code{\link{et0_tabari_talee1}()},
\code{\link{et0_tabari_talee2}()}
}
\concept{temperature models}
