% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0_temperature.R
\name{et0_droogers_allen}
\alias{et0_droogers_allen}
\title{Droogers-Allen ET0 Model}
\usage{
et0_droogers_allen(Tmin, Tmax, J, lat)
}
\arguments{
\item{Tmin}{Minimum air temperature (degrees Celsius).}

\item{Tmax}{Maximum air temperature (degrees Celsius).}

\item{J}{Day of the year (1-366).}

\item{lat}{Latitude (degrees).}
}
\value{
Daily ET0 (mm/day).
}
\description{
Estimates daily ET0 using the Droogers and Allen (2002) method, a modified
Hargreaves formula using the temperature range raised to the power 0.4.
}
\examples{
et0_droogers_allen(15, 30, 172, 25.43)
}
\references{
Droogers, P. & Allen, R.G. (2002). Estimating reference
  evapotranspiration under inaccurate data conditions. Irrigation and
  Drainage Systems, 16(1), 33-45.
}
\seealso{
Other temperature models: 
\code{\link{et0_baier_robertson}()},
\code{\link{et0_berti}()},
\code{\link{et0_blaney_criddle}()},
\code{\link{et0_dorji}()},
\code{\link{et0_hargreaves_samani}()},
\code{\link{et0_linacre}()},
\code{\link{et0_schendel}()},
\code{\link{et0_tabari_talee1}()},
\code{\link{et0_tabari_talee2}()}
}
\concept{temperature models}
