% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0_fao_pm.R
\name{et0_fao_pm}
\alias{et0_fao_pm}
\title{FAO Penman-Monteith Reference Evapotranspiration}
\usage{
et0_fao_pm(Tmin, Tmax, RH_morning, RH_evening, u2, n, J, lat, z = 0)
}
\arguments{
\item{Tmin}{Minimum air temperature (degrees Celsius).}

\item{Tmax}{Maximum air temperature (degrees Celsius).}

\item{RH_morning}{Morning relative humidity (percent).}

\item{RH_evening}{Evening relative humidity (percent).}

\item{u2}{Wind speed at 2 m height (m/s).}

\item{n}{Actual sunshine hours.}

\item{J}{Day of the year (1-366).}

\item{lat}{Latitude (degrees, negative for Southern Hemisphere).}

\item{z}{Elevation above sea level (m).}
}
\value{
Daily ET0 (mm/day).
}
\description{
Computes daily reference evapotranspiration (ET0) using the FAO
Penman-Monteith equation (Allen et al., 1998). This is the standard
reference method recommended by FAO.
}
\examples{
et0_fao_pm(Tmin = 15, Tmax = 30, RH_morning = 90, RH_evening = 40,
           u2 = 1.5, n = 8, J = 172, lat = 25.43, z = 216)
}
\references{
Allen, R.G., Pereira, L.S., Raes, D., & Smith, M. (1998). Crop
evapotranspiration: Guidelines for computing crop water requirements.
FAO Irrigation and Drainage Paper No. 56.
}
