% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0_temperature.R
\name{et0_hargreaves_samani}
\alias{et0_hargreaves_samani}
\title{Hargreaves-Samani ET0 Model}
\usage{
et0_hargreaves_samani(Tmin, Tmax, J, lat)
}
\arguments{
\item{Tmin}{Minimum air temperature (degrees Celsius).}

\item{Tmax}{Maximum air temperature (degrees Celsius).}

\item{J}{Day of the year (1-366).}

\item{lat}{Latitude (degrees).}
}
\value{
Daily ET0 (mm/day).
}
\description{
Estimates daily ET0 using the Hargreaves-Samani (1985) method based on
temperature range and extraterrestrial radiation.
}
\examples{
et0_hargreaves_samani(15, 30, 172, 25.43)
}
\references{
Hargreaves, G.H. & Samani, Z.A. (1985). Reference crop
  evapotranspiration from temperature. Applied Engineering in Agriculture,
  1(2), 96-99.
}
\seealso{
Other temperature models: 
\code{\link{et0_baier_robertson}()},
\code{\link{et0_berti}()},
\code{\link{et0_blaney_criddle}()},
\code{\link{et0_dorji}()},
\code{\link{et0_droogers_allen}()},
\code{\link{et0_linacre}()},
\code{\link{et0_schendel}()},
\code{\link{et0_tabari_talee1}()},
\code{\link{et0_tabari_talee2}()}
}
\concept{temperature models}
