% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0_temperature.R
\name{et0_linacre}
\alias{et0_linacre}
\title{Linacre ET0 Model}
\usage{
et0_linacre(Tmean, Tmin, Tmax, RH_morning, RH_evening, z, lat)
}
\arguments{
\item{Tmean}{Mean air temperature (degrees Celsius).}

\item{Tmin}{Minimum air temperature (degrees Celsius).}

\item{Tmax}{Maximum air temperature (degrees Celsius).}

\item{RH_morning}{Morning relative humidity (percent).}

\item{RH_evening}{Evening relative humidity (percent).}

\item{z}{Elevation above sea level (m).}

\item{lat}{Latitude (degrees).}
}
\value{
Daily ET0 (mm/day).
}
\description{
Estimates daily ET0 using the Linacre (1977) method, which accounts for
altitude, latitude and dew point depression.
}
\examples{
et0_linacre(25, 15, 35, 90, 40, 216, 25.43)
}
\references{
Linacre, E.T. (1977). A simple formula for estimating
  evaporation rates in various climates, using temperature data alone.
  Agricultural Meteorology, 18(6), 409-424.
}
\seealso{
Other temperature models: 
\code{\link{et0_baier_robertson}()},
\code{\link{et0_berti}()},
\code{\link{et0_blaney_criddle}()},
\code{\link{et0_dorji}()},
\code{\link{et0_droogers_allen}()},
\code{\link{et0_hargreaves_samani}()},
\code{\link{et0_schendel}()},
\code{\link{et0_tabari_talee1}()},
\code{\link{et0_tabari_talee2}()}
}
\concept{temperature models}
