% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0_temperature.R
\name{et0_schendel}
\alias{et0_schendel}
\title{Schendel ET0 Model}
\usage{
et0_schendel(Tmean, RH)
}
\arguments{
\item{Tmean}{Mean air temperature (degrees Celsius).}

\item{RH}{Mean relative humidity (percent).}
}
\value{
Daily ET0 (mm/day).
}
\description{
Estimates daily ET0 using the Schendel method (Bormann, 2011), which uses
mean temperature and relative humidity.
}
\examples{
et0_schendel(25, 60)
}
\references{
Bormann, H. (2011). Sensitivity analysis of 18 different
  potential evapotranspiration models to observed climatic change at German
  climate stations. Climatic Change, 104(3), 729-753.
}
\seealso{
Other temperature models: 
\code{\link{et0_baier_robertson}()},
\code{\link{et0_berti}()},
\code{\link{et0_blaney_criddle}()},
\code{\link{et0_dorji}()},
\code{\link{et0_droogers_allen}()},
\code{\link{et0_hargreaves_samani}()},
\code{\link{et0_linacre}()},
\code{\link{et0_tabari_talee1}()},
\code{\link{et0_tabari_talee2}()}
}
\concept{temperature models}
