% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0_temperature.R
\name{et0_tabari_talee2}
\alias{et0_tabari_talee2}
\title{Tabari-Talaee Model 2 ET0}
\usage{
et0_tabari_talee2(Tmin, Tmax, J, lat)
}
\arguments{
\item{Tmin}{Minimum air temperature (degrees Celsius).}

\item{Tmax}{Maximum air temperature (degrees Celsius).}

\item{J}{Day of the year (1-366).}

\item{lat}{Latitude (degrees).}
}
\value{
Daily ET0 (mm/day).
}
\description{
Estimates daily ET0 using the second Tabari and Talaee (2011) modified
Hargreaves equation with coefficient 0.0028.
}
\examples{
et0_tabari_talee2(15, 30, 172, 25.43)
}
\references{
Tabari, H. & Talaee, P.H. (2011). Local calibration of the
  Hargreaves and Priestley-Taylor equations for estimating reference
  evapotranspiration in arid and cold climates of Iran based on the
  Penman-Monteith model. Journal of Hydrologic Engineering, 16(10), 837-845.
}
\seealso{
Other temperature models: 
\code{\link{et0_baier_robertson}()},
\code{\link{et0_berti}()},
\code{\link{et0_blaney_criddle}()},
\code{\link{et0_dorji}()},
\code{\link{et0_droogers_allen}()},
\code{\link{et0_hargreaves_samani}()},
\code{\link{et0_linacre}()},
\code{\link{et0_schendel}()},
\code{\link{et0_tabari_talee1}()}
}
\concept{temperature models}
