% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extraterrestrial_radiation}
\alias{extraterrestrial_radiation}
\title{Extraterrestrial Radiation}
\usage{
extraterrestrial_radiation(J, lat)
}
\arguments{
\item{J}{Day of the year (1-366).}

\item{lat}{Latitude (degrees, negative for Southern Hemisphere).}
}
\value{
Extraterrestrial radiation (MJ/m2/day).
}
\description{
Computes the extraterrestrial radiation (Ra) for a given day and latitude
following Allen et al. (1998).
}
\examples{
extraterrestrial_radiation(172, 25.43)
}
\references{
Allen et al. (1998), FAO-56, Eq. 21.
}
