% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{kpa_to_hpa}
\alias{kpa_to_hpa}
\title{Convert Vapor Pressure from kPa to hPa}
\usage{
kpa_to_hpa(kPa)
}
\arguments{
\item{kPa}{Vapor pressure in kPa.}
}
\value{
Vapor pressure in hPa.
}
\description{
Convert Vapor Pressure from kPa to hPa
}
\keyword{internal}
