% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{kpa_to_mmhg}
\alias{kpa_to_mmhg}
\title{Convert Vapor Pressure from kPa to mmHg}
\usage{
kpa_to_mmhg(kPa)
}
\arguments{
\item{kPa}{Vapor pressure in kPa.}
}
\value{
Vapor pressure in mmHg.
}
\description{
Convert Vapor Pressure from kPa to mmHg
}
\keyword{internal}
