% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{latent_heat}
\alias{latent_heat}
\title{Latent Heat of Vaporization}
\usage{
latent_heat(Tmean)
}
\arguments{
\item{Tmean}{Mean air temperature (degrees Celsius).}
}
\value{
Latent heat of vaporization (MJ/kg).
}
\description{
Computes the latent heat of vaporization as a function of air temperature
following Allen et al. (1998).
}
\examples{
latent_heat(25)
}
\references{
Allen, R.G., Pereira, L.S., Raes, D., & Smith, M. (1998).
  FAO Irrigation and Drainage Paper No. 56.
}
