% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ms_to_miles_day}
\alias{ms_to_miles_day}
\title{Convert Wind Speed from m/s to miles/day}
\usage{
ms_to_miles_day(u)
}
\arguments{
\item{u}{Wind speed in m/s.}
}
\value{
Wind speed in miles/day.
}
\description{
Convert Wind Speed from m/s to miles/day
}
\keyword{internal}
