% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{net_radiation}
\alias{net_radiation}
\title{Net Radiation}
\usage{
net_radiation(J, lat, n, Tmin, Tmax, ea, albedo = 0.23, a_s = 0.25, b_s = 0.5)
}
\arguments{
\item{J}{Day of the year (1-366).}

\item{lat}{Latitude (degrees).}

\item{n}{Actual sunshine hours.}

\item{Tmin}{Minimum air temperature (degrees Celsius).}

\item{Tmax}{Maximum air temperature (degrees Celsius).}

\item{ea}{Actual vapor pressure (kPa).}

\item{albedo}{Surface albedo (default 0.23 for reference crop).}

\item{a_s}{Angstrom coefficient a (default 0.25).}

\item{b_s}{Angstrom coefficient b (default 0.50).}
}
\value{
Net radiation (MJ/m2/day).
}
\description{
Computes net radiation (Rn) following the FAO-56 approach.
}
\examples{
ea <- actual_vapor_pressure(15, 30, 90, 40)
net_radiation(172, 25.43, 8, 15, 30, ea)
}
\references{
Allen et al. (1998), FAO-56.
}
