% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{psychrometric_constant}
\alias{psychrometric_constant}
\title{Psychrometric Constant}
\usage{
psychrometric_constant(P = NULL, z = 0)
}
\arguments{
\item{P}{Atmospheric pressure (kPa). If \code{NULL}, computed from elevation.}

\item{z}{Elevation above sea level (m). Used only if \code{P} is \code{NULL}.}
}
\value{
Psychrometric constant (kPa/degrees C).
}
\description{
Computes the psychrometric constant as a function of atmospheric pressure.
}
\examples{
psychrometric_constant(z = 216)
}
\references{
Allen et al. (1998), FAO-56, Eq. 8.
}
