% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{slope_vapor_pressure}
\alias{slope_vapor_pressure}
\title{Slope of Saturation Vapor Pressure Curve}
\usage{
slope_vapor_pressure(Tmean)
}
\arguments{
\item{Tmean}{Mean air temperature (degrees Celsius).}
}
\value{
Slope of saturation vapor pressure curve (kPa/degrees C).
}
\description{
Computes the slope of the saturation vapor pressure-temperature curve.
}
\examples{
slope_vapor_pressure(25)
}
\references{
Allen et al. (1998), FAO-56, Eq. 13.
}
