% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0_temperature.R
\name{temperature_models}
\alias{temperature_models}
\title{Temperature-Based ET0 Models}
\description{
Functions implementing 10 temperature-based empirical models
  for reference evapotranspiration (ET0) estimation.
}
