% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{saturation_vapor_pressure}
\alias{saturation_vapor_pressure}
\title{Saturation Vapor Pressure}
\usage{
saturation_vapor_pressure(T)
}
\arguments{
\item{T}{Air temperature (degrees Celsius).}
}
\value{
Saturation vapor pressure (kPa).
}
\description{
Computes the saturation vapor pressure at a given temperature using the
Tetens formula (Allen et al., 1998).
}
\examples{
saturation_vapor_pressure(25)
}
\references{
Allen et al. (1998), FAO-56, Eq. 11.
}
